added home page and updated navigator structure
